# Removes wasm.FunctionInstance type