--- a/internal/engine/compiler/arch.go
+++ b/internal/engine/compiler/arch.go
@@ -1,6 +1,9 @@
 package compiler
 
-import "github.com/tetratelabs/wazero/internal/asm"
+import (
+	"github.com/tetratelabs/wazero/internal/asm"
+	"github.com/tetratelabs/wazero/internal/wasm"
+)
 
 // newArchContext returns a new archContext which is architecture-specific type to be embedded in callEngine.
 // This must be initialized in init() function in architecture-specific arch_*.go file which is guarded by build tag.
@@ -11,7 +14,7 @@ var newArchContext func() archContext
 // ce is "*callEngine" as uintptr.
 //
 // Note: this is implemented in per-arch Go assembler file. For example, arch_amd64.s implements this for amd64.
-func nativecall(codeSegment, ce uintptr, moduleInstanceAddress uintptr)
+func nativecall(codeSegment, ce uintptr, moduleInstanceAddress *wasm.ModuleInstance)
 
 // registerNameFn is used for debugging purpose to have register symbols in the string of runtimeValueLocation.
 var registerNameFn func(register asm.Register) string