… internal/engine/compiler/compiler_bench_test.go
… internal/engine/compiler/compiler_controlflow_test.go
… internal/engine/compiler/compiler_initialization_test.go
… internal/engine/compiler/compiler_memory_test.go
… internal/engine/compiler/compiler_post1_0_test.go
… internal/engine/compiler/compiler_test.go
… internal/engine/compiler/engine_test.go
… internal/engine/compiler/impl_amd64_test.go
… internal/engine/compiler/impl_arm64_test.go
… internal/engine/interpreter/interpreter_test.go
… internal/integration_test/bench/hostfunc_bench_test.go
… internal/wasm/function_definition.go
… internal/wasm/function_definition_test.go
… internal/wasm/module.go
… internal/wasm/module_instance.go
… internal/wasm/module_test.go
… internal/wasm/store.go
… internal/wasm/store_test.go
… runtime_test.go