FunctionInstance has existed for historical reasons, but now it's unnecessary due to
the existence of Engine specific `function`. This not only simplifies the logic, but 
also reduces the memory usage during instantiation.


```
$ benchstat old.txt new.txt
goos: linux
goarch: amd64
pkg: github.com/tetratelabs/wazero/internal/integration_test/bench
cpu: AMD Ryzen 9 3950X 16-Core Processor
                                       │   old.txt   │              new.txt               │
                                       │   sec/op    │   sec/op     vs base               │
Initialization/interpreter-32            32.65µ ± 0%   32.76µ ± 1%       ~ (p=0.665 n=30)
Initialization/interpreter-multiple-32   14.15µ ± 2%   14.42µ ± 1%  +1.90% (p=0.000 n=30)
Initialization/compiler-32               38.20µ ± 1%   38.05µ ± 1%       ~ (p=0.311 n=30)
Initialization/compiler-multiple-32      13.64µ ± 1%   13.71µ ± 1%       ~ (p=0.171 n=30)
Compilation/with_extern_cache-32         288.3µ ± 1%   289.9µ ± 1%       ~ (p=0.093 n=30)
Compilation/without_extern_cache-32      10.51m ± 1%   10.58m ± 1%       ~ (p=0.142 n=30)
geomean                                  94.88µ        95.43µ       +0.58%

                                       │   old.txt    │               new.txt               │
                                       │     B/op     │     B/op      vs base               │
Initialization/interpreter-32            132.5Ki ± 0%   132.6Ki ± 0%  +0.09% (p=0.000 n=30)
Initialization/interpreter-multiple-32   132.9Ki ± 0%   133.0Ki ± 0%  +0.09% (p=0.000 n=30)
Initialization/compiler-32               133.1Ki ± 0%   132.6Ki ± 0%  -0.41% (p=0.000 n=30)
Initialization/compiler-multiple-32      137.4Ki ± 0%   136.8Ki ± 0%  -0.40% (p=0.000 n=30)
Compilation/with_extern_cache-32         53.81Ki ± 0%   53.75Ki ± 0%  -0.10% (p=0.000 n=30)
Compilation/without_extern_cache-32      1.907Mi ± 0%   1.907Mi ± 0%  +0.01% (p=0.023 n=30)
geomean                                  179.8Ki        179.6Ki       -0.12%

                                       │   old.txt   │              new.txt               │
                                       │  allocs/op  │  allocs/op   vs base               │
Initialization/interpreter-32             28.00 ± 0%    27.00 ± 0%  -3.57% (p=0.000 n=30)
Initialization/interpreter-multiple-32    41.00 ± 0%    40.00 ± 0%  -2.44% (p=0.000 n=30)
Initialization/compiler-32                28.00 ± 0%    27.00 ± 0%  -3.57% (p=0.000 n=30)
Initialization/compiler-multiple-32       31.00 ± 0%    30.00 ± 0%  -3.23% (p=0.000 n=30)
Compilation/with_extern_cache-32          986.0 ± 0%    986.0 ± 0%       ~ (p=0.313 n=30)
Compilation/without_extern_cache-32      32.62k ± 0%   32.67k ± 0%  +0.15% (p=0.000 n=30)
geomean                                   178.2         174.4       -2.12%

```